// supabase/functions/fetch-recently-listened/spotify/types.ts
export type FetchJSONResponse<T> = T & {
  error?: Record<string, any>;
};
// ...
type TrackObject = {
  album?: Album;
  artists?: ArtistObject[];
  available_markets?: string[];
  disc_number?: number;
  duration_ms?: number;
  explicit?: boolean;
  external_ids?: {
    isrc?: string;
    ean?: string;
    upc?: string;
  };
  external_urls?: {
    spotify?: string;
  };
  href?: string;
  id?: string;
  is_playable?: boolean;
  linked_from?: Record<string, never>;
  restrictions?: {
    reason: 'market' | 'product' | 'explicit';
  };
  name?: string;
  popularity?: number;
  preview_url?: string | null;
  track_number?: number;
  type?: 'track';
  uri?: string;
  is_local?: boolean;
};
// ...
type PlayHistoryObject = {
  track: TrackObject;
  played_at: string;
  context: {
    type: string;
    href: string;
    external_urls: {
      spotify: string;
    };
    uri: string;
  };
};

export type RecentlyListenedResponse = {
  href: string;
  limit: number;
  next: string | null;
  cursors: {
    after: string;
    before: string;
  };
  total: number;
  items: PlayHistoryObject[];
};